move TestRealDenoCacheEnv into library
