+++ rs_lib/src/global/cache_file.rs
@@ -1,3 +1,5 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
 use std::io::ErrorKind;
 use std::path::Path;
